requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5

# Authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Email
Flask-Mail==0.9.1

# Real-time (Chat)
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.33.3

# Azure
azure-storage-blob==12.19.0

# Security
python-dotenv==1.0.0
Flask-Limiter==3.5.0

# Utils
Pillow==10.1.0
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0

